Create MQ9.ino
This code is used to detect and measure the presence of smoke in any facility. Although this code was designed for the MQ-9 gas sensor, any of the MQ sensors can be used with suitable modifications to the threshold value of the "ppm" variable (which is a measure of sensor output).  
You have to connect the output of the gas sensor to the A0 pin in the NodeMCU.